Convert InfoMessage to a reusable component
